Rename query-runner to new-query-runner to match class and file
